Pages Restricted according to Authorization

+++ backend/api/views.py
@@ -1,12 +1,18 @@
 from django.shortcuts import render, redirect
+from django.http import HttpResponse
 from django.db.models import Q
 from django.contrib.auth.models import User
 from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.decorators import login_required
 from django.contrib import messages
 from .models import Room, Topic
 from .forms import RoomForm
 
 def login_portal(request):
+    # if user is already logged in and tries to route the `/login`, the user will be redirected to the hompage.
+    if request.user.is_authenticated:
+        return redirect('homepage')
+
     if request.method == 'POST':
         username = request.POST.get('username')
         password = request.POST.get('password')
@@ -45,11 +51,13 @@ def home(request):
     context = {'rooms': rooms, 'topics': topics, 'room_count': room_count}
     return render(request, 'api/home.html', context)
 
+@login_required(login_url='/login')
 def room(request, pk):
     room = Room.objects.get(id=pk)    
     context = {'room': room}
     return render(request, 'api/room.html', context)
 
+@login_required(login_url='/login')
 def create_room(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -61,9 +69,14 @@ def create_room(request):
     context = {'form': form}
     return render(request, 'api/room_form.html', context)
 
+@login_required(login_url='/login')
 def update_room(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse('You are not allowed to update this Room!!')
+
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
         if form.is_valid():
@@ -73,8 +86,13 @@ def update_room(request, pk):
     context = {'form': form}
     return render(request, 'api/room_form.html', context)
 
+@login_required(login_url='/login')
 def delete_room(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse('You are not allowed to delete this Room!!')
+    
     if request.method == 'POST':
         room.delete()
         return redirect('homepage')
